Add files via upload
The business objectives of this project are as follows:
Develop the prototype of Employee Management System for Conference center at Mozingo lake recreation park which can be effectively implemented by our client at minimal cost and utilizing less number of resources. This project follows Agile scrum process model or methodology for the development of the project. This scrum model uses a product backlog, a prioritized list of high-level requirements written as User Stories. At the start of a sprint, many user stories are selected and put into the Sprint Backlog. These stories are the features that will be implemented during the sprint. For each user story different phases are traversed: requirements, Use case analysis, System Model, Data Model, User Interface Design implementation and testing. Each sprint a duration. After the sprints, the product must be delivered to the customer.

My responsibilities in this project:
•	Functioned using an Agile Scrum methodology to design a Database Structure and a Prototype for an Employee Management System, for our clients being the officials of Mozingo Lake Recreation Park, City of Maryville, MO
•	Gathered requirements during the inception phase; documented and delivered, the system and project management documents
•	Assisted in managing the product and sprint backlogs; also aided to develop the System and Data model
•	Designed the wireframes or prototype screens for the system; helped the client with design walkthrough
•	Rendered Team and Project Management skills by applying theoretical knowledge into practice while working for real-time clients